2011-08-19  Chris Sutcliffe  <ir0nh34d@users.sourceforge.net>
        * include/sys/types.h (ssize_t): Defined as int as opposed to long.

        Thanks to bvassche for the report.

+++ winsup/mingw/include/sys/types.h
@@ -112,7 +112,7 @@ typedef _sigset_t	sigset_t;
 
 #ifndef _SSIZE_T_
 #define _SSIZE_T_
-typedef long _ssize_t;
+typedef int _ssize_t;
 
 #ifndef	_NO_OLDNAMES
 typedef _ssize_t ssize_t;
